refactor(evaluatorq-py): import TokenUsage from contracts, not simulation.types (RES-897)
